app/requirement.txt


Flask==2.3.2
mysql-connector-python==8.0.33
gunicorn==21.2.0

